fix: limit concurrent HTTP requests
All HTTP requests made by this module are sent to the same delegate
host. Browsers throttle the number of concurrent requests per hostname,
right now it is 6 per host, which suffocates the use of delegate and
blocking it from being used for preload or delegated content routing.

This introduces a task queue that limits the number of concurrent
requests, making it safe to run in browser context.

See also:
https://github.com/libp2p/js-libp2p-delegated-content-routing/issues/12

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>